Refactored gameloop
- game.js holds game components
- index.js starts the game
