+++ lib/index.js
@@ -1,24 +1,3 @@
-var KeyListener = require("./keyListener");
-var Player = require("./player");
+var Game = require("./game");
 
-var canvas  = document.getElementById("screen");
-var ctx = canvas.getContext("2d");
-
-// initialize the game
-  // create players
-  // start listeners
-  // start gameLoop
-
-// gameLoop
-
-canvas.focus();
-
-KeyListener.start();
-
-var firstPlayer = new Player(50, 50, 10, 10);
-
-requestAnimationFrame(function gameLoop() {
-  firstPlayer.move();
-  ctx.fillRect(firstPlayer.x, firstPlayer.y, firstPlayer.width, firstPlayer.height);
-  requestAnimationFrame(gameLoop);
-});
+new Game();
